refactor(isHref): rename isHref file to isLinkTarget
