+++ packages/admin/cms-admin/src/blocks/ExternalLinkBlock.tsx
@@ -5,7 +5,7 @@ import * as React from "react";
 import { FormattedMessage } from "react-intl";
 
 import { ExternalLinkBlockData, ExternalLinkBlockInput } from "../blocks.generated";
-import { isLinkTarget } from "../validation/isHref";
+import { isLinkTarget } from "../validation/isLinkTarget";
 import { validateUrl } from "../validation/validateUrl";
 
 type State = ExternalLinkBlockData;
